// src/fm-v1.8/types/ethers-contracts/Size.ts
export type BuyCreditLimitParamsStruct = {
  maturities: BigNumberish[];
  aprs: BigNumberish[];
};

export type BuyCreditLimitParamsStructOutput = [BigNumber[], BigNumber[]] & {
  maturities: BigNumber[];
  aprs: BigNumber[];
};

export type BuyCreditLimitOnBehalfOfParamsStruct = {
  params: BuyCreditLimitParamsStruct;
  onBehalfOf: string;
};

export type BuyCreditLimitOnBehalfOfParamsStructOutput = [
  BuyCreditLimitParamsStructOutput,
  string
] & { params: BuyCreditLimitParamsStructOutput; onBehalfOf: string };

export type BuyCreditMarketParamsStruct = {
  borrower: string;
  creditPositionId: BigNumberish;
  amount: BigNumberish;
  maturity: BigNumberish;
  deadline: BigNumberish;
  minAPR: BigNumberish;
  exactAmountIn: boolean;
  collectionId: BigNumberish;
  rateProvider: string;
};

export type BuyCreditMarketParamsStructOutput = [
  string,
  BigNumber,
  BigNumber,
  BigNumber,
  BigNumber,
  BigNumber,
  boolean,
  BigNumber,
  string
] & {
  borrower: string;
  creditPositionId: BigNumber;
  amount: BigNumber;
  maturity: BigNumber;
  deadline: BigNumber;
  minAPR: BigNumber;
  exactAmountIn: boolean;
  collectionId: BigNumber;
  rateProvider: string;
};

export type BuyCreditMarketOnBehalfOfParamsStruct = {
  params: BuyCreditMarketParamsStruct;
  onBehalfOf: string;
  recipient: string;
};
// ...
export type DepositParamsStruct = {
  token: string;
  amount: BigNumberish;
  to: string;
};

export type DepositParamsStructOutput = [string, BigNumber, string] & {
  token: string;
  amount: BigNumber;
  to: string;
};

export type DepositOnBehalfOfParamsStruct = {
  params: DepositParamsStruct;
  onBehalfOf: string;
};
// ...
export type SellCreditMarketParamsStruct = {
  lender: string;
  creditPositionId: BigNumberish;
  amount: BigNumberish;
  maturity: BigNumberish;
  deadline: BigNumberish;
  maxAPR: BigNumberish;
  exactAmountIn: boolean;
  collectionId: BigNumberish;
  rateProvider: string;
};

export type SellCreditMarketParamsStructOutput = [
  string,
  BigNumber,
  BigNumber,
  BigNumber,
  BigNumber,
  BigNumber,
  boolean,
  BigNumber,
  string
] & {
  lender: string;
  creditPositionId: BigNumber;
  amount: BigNumber;
  maturity: BigNumber;
  deadline: BigNumber;
  maxAPR: BigNumber;
  exactAmountIn: boolean;
  collectionId: BigNumber;
  rateProvider: string;
};

export type CopyLimitOrderConfigStruct = {
  minTenor: BigNumberish;
  maxTenor: BigNumberish;
  minAPR: BigNumberish;
  maxAPR: BigNumberish;
  offsetAPR: BigNumberish;
};
// ...
export type SelfLiquidateParamsStruct = { creditPositionId: BigNumberish };

export type SelfLiquidateParamsStructOutput = [BigNumber] & {
  creditPositionId: BigNumber;
};

export type SelfLiquidateOnBehalfOfParamsStruct = {
  params: SelfLiquidateParamsStruct;
  onBehalfOf: string;
  recipient: string;
};

export type SelfLiquidateOnBehalfOfParamsStructOutput = [
  SelfLiquidateParamsStructOutput,
  string,
  string
] & {
  params: SelfLiquidateParamsStructOutput;
  onBehalfOf: string;
  recipient: string;
};

export type SellCreditLimitParamsStruct = {
  maturities: BigNumberish[];
  aprs: BigNumberish[];
};

export type SellCreditLimitParamsStructOutput = [BigNumber[], BigNumber[]] & {
  maturities: BigNumber[];
  aprs: BigNumber[];
};

export type SellCreditLimitOnBehalfOfParamsStruct = {
  params: SellCreditLimitParamsStruct;
  onBehalfOf: string;
};

export type SellCreditLimitOnBehalfOfParamsStructOutput = [
  SellCreditLimitParamsStructOutput,
  string
] & { params: SellCreditLimitParamsStructOutput; onBehalfOf: string };

export type SellCreditMarketOnBehalfOfParamsStruct = {
  params: SellCreditMarketParamsStruct;
  onBehalfOf: string;
  recipient: string;
};

export type SellCreditMarketOnBehalfOfParamsStructOutput = [
  SellCreditMarketParamsStructOutput,
  string,
  string
] & {
  params: SellCreditMarketParamsStructOutput;
  onBehalfOf: string;
  recipient: string;
};

export type SetCopyLimitOrderConfigsParamsStruct = {
  copyLoanOfferConfig: CopyLimitOrderConfigStruct;
  copyBorrowOfferConfig: CopyLimitOrderConfigStruct;
};

export type SetCopyLimitOrderConfigsParamsStructOutput = [
  CopyLimitOrderConfigStructOutput,
  CopyLimitOrderConfigStructOutput
] & {
  copyLoanOfferConfig: CopyLimitOrderConfigStructOutput;
  copyBorrowOfferConfig: CopyLimitOrderConfigStructOutput;
};

export type SetCopyLimitOrderConfigsOnBehalfOfParamsStruct = {
  params: SetCopyLimitOrderConfigsParamsStruct;
  onBehalfOf: string;
};

export type SetCopyLimitOrderConfigsOnBehalfOfParamsStructOutput = [
  SetCopyLimitOrderConfigsParamsStructOutput,
  string
] & { params: SetCopyLimitOrderConfigsParamsStructOutput; onBehalfOf: string };

export type SetUserConfigurationParamsStruct = {
  openingLimitBorrowCR: BigNumberish;
  allCreditPositionsForSaleDisabled: boolean;
  creditPositionIdsForSale: boolean;
  creditPositionIds: BigNumberish[];
};

export type SetUserConfigurationParamsStructOutput = [
  BigNumber,
  boolean,
  boolean,
  BigNumber[]
] & {
  openingLimitBorrowCR: BigNumber;
  allCreditPositionsForSaleDisabled: boolean;
  creditPositionIdsForSale: boolean;
  creditPositionIds: BigNumber[];
};

export type SetUserConfigurationOnBehalfOfParamsStruct = {
  params: SetUserConfigurationParamsStruct;
  onBehalfOf: string;
};

export type SetUserConfigurationOnBehalfOfParamsStructOutput = [
  SetUserConfigurationParamsStructOutput,
  string
] & { params: SetUserConfigurationParamsStructOutput; onBehalfOf: string };

export type SetVaultParamsStruct = { vault: string; forfeitOldShares: boolean };

export type SetVaultParamsStructOutput = [string, boolean] & {
  vault: string;
  forfeitOldShares: boolean;
};

export type SetVaultOnBehalfOfParamsStruct = {
  params: SetVaultParamsStruct;
  onBehalfOf: string;
};

export type SetVaultOnBehalfOfParamsStructOutput = [
  SetVaultParamsStructOutput,
  string
] & { params: SetVaultParamsStructOutput; onBehalfOf: string };

export type UpdateConfigParamsStruct = { key: string; value: BigNumberish };

export type UpdateConfigParamsStructOutput = [string, BigNumber] & {
  key: string;
  value: BigNumber;
};

export type WithdrawParamsStruct = {
  token: string;
  amount: BigNumberish;
  to: string;
};

export type WithdrawParamsStructOutput = [string, BigNumber, string] & {
  token: string;
  amount: BigNumber;
  to: string;
};

export type WithdrawOnBehalfOfParamsStruct = {
  params: WithdrawParamsStruct;
  onBehalfOf: string;
};